% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATCosinor2.R
\name{CATCosinor2}
\alias{CATCosinor2}
\title{Cosinor Analysis 2.0}
\usage{
CATCosinor2(data, TimeCol = 1, Y = 2, Components = 1,
  window = "noTaper", RefDateTime = NA, timeFormat = "\%Y\%m\%d\%H\%M",
  RangeDateTime = list(Start = NA, End = NA), Units = "hours", dt = 0,
  Progressive = list(Interval = 0, Increment = 0), Period = list(Set = 0,
  Start = 0, Increment = 1, End = 0), Debug = FALSE, IDcol = "fileName",
  GrpCol = NA)
}
\arguments{
\item{TimeCol}{Column(s) where time is found.  c(1,2) can be used when date is in column 1 and time is in column 2.
Date and time can be in any two columns, i.e., c(6,3). If time is in two columns, the formats should be \%d/\%m/\%y 
in the first of the two columns, and \%H:\%M:\%S in the second of the two (Any time format in timeFormat paramater is 
ignored in this case.)}

\item{Y}{Column holding data to be analyzed}

\item{Components}{Default=1.  Indicates if this is a single or multiple component cosinor analysis, where the number of components is specified (>0).}

\item{window}{Data will be convolved with a window:  Hanning, Hamming, Bartlett, Blackman.  Default="noTaper"}

\item{RefDateTime}{Default=NA  Date used as reference, and subtracted from all data dates, to make the number smaller. 
The format is always "\%Y\%m\%d\%H\%M" not implemented---if RefDateTime = NA, use the 1st date of the data as the RefDateTime
if RefDateTime = 0, use midnight of the same day as the data starts when using years with leading 0s  (0500 AD) be 
sure to include the leading 0, and put quotes around the RefDateTime}

\item{timeFormat}{RefDateTime and the time in the data file should both use the same date/time format.
Can be "numeric" or "\%Y\%m\%d\%H\%M"??? i.e., "\%Y\%m\%d\%H\%M", or if = "numeric", time column in data file can be 
simple numbers (0 - 99999...) if "numeric", data is sorted by time to be sure they are ordered ascending.  
First must be smallest , and last largest. Time can also be in two columns (indicate in TimeCol);  
timeFormat is ignored when time is in two columns -- the format use is \%d/\%m/\%y in the first of the two 
columns, and \%H:\%M:\%S or \%H:\%M in the second of the two}

\item{Units}{Units (hour, year, week or day) for Interval and Increment arguments, as well as 
Period arguments, and any time variable. ToDo:  modify to use units other than hours -- currently only works for hours}

\item{dt}{When equidistant data, dt indicates the sampling interval.  If dt =0, no periodogram is done.
Data is assumed to be equidistant when this is nonzero}
}

