\name{CATkit-package}
\alias{CATkit}
\docType{package}
\title{
Chronomics Analysis Toolkit (CAT):  Periodicity Analysis
}
\description{
http://564394709114639785.weebly.com
This package contains functions that can be used to identify periodicities in time series.  The two calls are: CATCosinor and CATCall.
CATCosinor runs various forms of a cosinor technique: a regression using cosine curves to test for anticipated periodic components and to estimate rhythm parameters: MESOR (rhythm-adjusted mean value), amplitude, and acrophase (phase of maximum).
CatCall can run the following time series analysis functions:  Smooth, Actogram, AutoCorr, CrossCorr.
}
\details{Detailed installation instructions and sample runs (vignettes) can be found at 
http://564394709114639785.weebly.com/installing-cat.html
}
\author{
Cathy Lee Gierke, Ruth Helget, Germaine Cornelissen-Guillaume

Maintainer: \email{ leegi001@umn.edu}

}
\keyword{ chronomics cosinor CAT Autocorrelation Crosscorrelation Actogram Smoothing }

