% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate CATCosinor Results}
\usage{
consolidate(path)
}
\arguments{
\item{path}{The path where the CATCosinor .dat output files reside}
}
\value{
A data.frame object
}
\description{
This function takes the path originally passed to
the CATCosinor function and collects all the relevant results
for subsequent analysis in a \code{data.frame} object.
}
