\name{CATkit-package}
\alias{CATkit}
\docType{package}
\title{
Chronomics Analysis Toolkit (CAT):  Analyze Periodicity
}
\description{
http://564394709114639785.weebly.com
This package contains functions that can be used to identify periodicities in time series data.  The two calls are:
CATCosinor runs various forms of a cosinor function.  This technique is a regression using cosine curves to identify the period, MESOR, amplitude and phase of sinusoidal rhythms in time series data.
CatCall can run the following time series analysis functions:  Smooth, Actogram, AutoCorr, CrossCorr.
}
\details{Detailed installation instructions and sample runs (vignettes) can be found at http://564394709114639785.weebly.com/installing-cat.html
}
\author{
Cathy Lee Gierke, Germaine Cornelissen-Guillaume, John A Lindgren

Maintainer: \email{ <leegi001@umn.edu>}

}
\keyword{ chronomics cosinor CAT Autocorrelation Crosscorrelation Actogram Smoothing }

